% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{download}
\alias{download}
\title{download jvms binary}
\usage{
download(path = jvms_loc(), force = F)
}
\arguments{
\item{path}{where to set jvms binary. Defualt is $HOME/.jvms}

\item{force}{download function check if already exist. So force is TRUE means ignore previous binary.}
}
\description{
download jvms binary
}
