% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exec.R
\name{jvms}
\alias{jvms}
\title{jvms commend}
\usage{
jvms(args = "", jvms = "auto")
}
\arguments{
\item{args}{jvms args.}

\item{jvms}{jvms binary location}
}
\description{
jvms commend
}
\examples{
  jvms()
  jvms("init")
}
